Refactoring in order to fix circular import issue
