+++ test_collections/matter/sdk_tests/support/python_testing/models/test_case.py
@@ -37,7 +37,8 @@ from ...pics import PICS_FILE_PATH
 from ...sdk_container import SDKContainer
 from ...utils import prompt_for_commissioning_mode
 from .python_test_models import PythonTest, PythonTestType
-from .python_testing_hooks_proxy import SDKPythonTestResultBase
+from ...sdk_utils.python_testing.python_testing_hooks_proxy import SDKPythonTestResultBase
+# from .python_testing_hooks_proxy import SDKPythonTestResultBase
 from .utils import (
     EXECUTABLE,
     RUNNER_CLASS_PATH,
